+++ Back-end/api-module/src/main/java/com/ico/api/service/StudentProductService.java
@@ -4,5 +4,5 @@ import com.ico.core.entity.StudentProduct;
 import org.springframework.web.bind.annotation.RequestParam;
 
 public interface StudentProductService {
-    void createProduct(String identity, Long nationId, String title, int amount, String image, String detail, int count);
+    void createProduct(String identity, Long nationId, String title, int amount, String image, String detail, byte count);
 }
